REPO-4791 Use Local transforms where Legacy transforms are called.
- Re factor TransformerDebug so the Legacy code has been extracted into LegacyTransformerDebug
  and code used by the Admin console's Test Transform that is not being deprecated
  has been extracted into AdminUiTransformerDebug
