+++ src/main/java/org/alfresco/repo/rendition2/LegacyTransformServiceRegistry.java
@@ -26,13 +26,13 @@
 package org.alfresco.repo.rendition2;
 
 import org.alfresco.repo.content.transform.TransformerDebug;
-import org.alfresco.service.cmr.repository.ContentService;
 import org.alfresco.service.cmr.repository.TransformationOptions;
 import org.alfresco.transform.client.registry.TransformServiceRegistry;
 import org.alfresco.util.PropertyCheck;
 import org.springframework.beans.factory.InitializingBean;
 
 import java.util.Map;
+
 
 /**
  * Implements {@link TransformServiceRegistry} providing a mechanism of validating if a legacy transformation
